Added to the MultipleChoice class and Country class
I realized the method findOtherCountries would only find another country other than the one in use but it could still have the same value if say the predominant language was the same as a bunch of other places; i.e. English, Spanish, French, etc.  I also added a portion to the Country class so that it would populate the Multiple Choice questions on its own.