Implemented CountryOfDay to display the country info
Made activity_country_of_day display CountryOfDay text, and made some instance variables in the country class final for better security and to satisfy the complaints of the compiler. The buttons still don't work though, and some work will need to be done making it pretty. Also, the CountryOfDay class will need to be updated to change the country every time a change in date is detected by the system